added deleteIssueById and deleteAllIssues API

+++ src/main/java/com/nubes/tms/service/IssueService.java
@@ -20,7 +20,7 @@ public interface IssueService {
 
 	List<Issue> getIssuesByPriority(String priority);
 
-	Integer deleteAllIssues();
+	Long deleteAllIssues();
 
 	boolean deleteIssue(String id);
